# sidecar/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "edgar-sidecar",
    "summary": "HTTP wrapper around edgartools - the single SEC egress",
    "version": "0.1.0"
  },
  "paths": {
...
    "/filing/{accession}": {
      "get": {
        "summary": "Get Filing",
        "operationId": "get_filing_filing__accession__get",
        "parameters": [
          {
            "name": "accession",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^\\d{10}-\\d{2}-\\d{6}$",
              "title": "Accession"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilingEnvelope"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          }
        }
      }
    },
    "/filing/{accession}/content": {
      "get": {
        "summary": "Get Filing Content",
        "operationId": "get_filing_content_filing__accession__content_get",
        "parameters": [
          {
            "name": "accession",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^\\d{10}-\\d{2}-\\d{6}$",
              "title": "Accession"
            }
          },
          {
            "name": "fmt",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "markdown",
                "text",
                "html"
              ],
              "type": "string",
              "default": "markdown",
              "title": "Fmt"
            }
          },
          {
            "name": "page_breaks",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Page Breaks"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          }
        }
      }
    },
    "/filing/{accession}/sections": {
      "get": {
        "summary": "Get Filing Sections",
        "operationId": "get_filing_sections_filing__accession__sections_get",
        "parameters": [
          {
            "name": "accession",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^\\d{10}-\\d{2}-\\d{6}$",
              "title": "Accession"
            }
          },
          {
            "name": "fmt",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "text",
                "markdown"
              ],
              "type": "string",
              "default": "text",
              "title": "Fmt"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SectionsResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          }
        }
      }
    },
    "/filing/{accession}/attachments": {
      "get": {
        "summary": "List Attachments",
        "operationId": "list_attachments_filing__accession__attachments_get",
        "parameters": [
          {
            "name": "accession",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^\\d{10}-\\d{2}-\\d{6}$",
              "title": "Accession"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachmentsResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          }
        }
      }
    },
    "/filing/{accession}/attachments/{sequence}": {
      "get": {
        "summary": "Get Attachment Content",
        "operationId": "get_attachment_content_filing__accession__attachments__sequence__get",
        "parameters": [
          {
            "name": "accession",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^\\d{10}-\\d{2}-\\d{6}$",
              "title": "Accession"
            }
          },
          {
            "name": "sequence",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^\\d{1,4}$",
              "title": "Sequence"
            }
          },
          {
            "name": "fmt",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "raw",
                "text",
                "markdown"
              ],
              "type": "string",
              "default": "text",
              "title": "Fmt"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachmentContentResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttachmentContentResponse": {
        "properties": {
          "accession_number": {
            "type": "string",
            "pattern": "^\\d{10}-\\d{2}-\\d{6}$",
            "title": "Accession Number"
          },
          "sequence": {
            "type": "string",
            "title": "Sequence"
          },
          "document": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document"
          },
          "document_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Type"
          },
          "fmt": {
            "type": "string",
            "enum": [
              "raw",
              "text",
              "markdown"
            ],
            "title": "Fmt"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "accession_number",
          "sequence",
          "document",
          "document_type",
          "fmt",
          "content"
        ],
        "title": "AttachmentContentResponse"
      },
      "AttachmentInfo": {
        "properties": {
          "sequence": {
            "type": "string",
            "title": "Sequence"
          },
          "document": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "display_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Description"
          },
          "purpose": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purpose"
          },
          "document_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Type"
          },
          "size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size"
          },
          "ixbrl": {
            "type": "boolean",
            "title": "Ixbrl"
          },
          "extension": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extension"
          },
          "is_binary": {
            "type": "boolean",
            "title": "Is Binary"
          },
          "is_primary": {
            "type": "boolean",
            "title": "Is Primary"
          },
          "group": {
            "type": "string",
            "enum": [
              "document",
              "data_file"
            ],
            "title": "Group"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sequence",
          "document",
          "description",
          "display_description",
          "purpose",
          "document_type",
          "size",
          "ixbrl",
          "extension",
          "is_binary",
          "is_primary",
          "group",
          "url"
        ],
        "title": "AttachmentInfo",
        "description": "One submission document; richer than the envelope's lean DocumentRef."
      },
      "AttachmentsResponse": {
        "properties": {
          "accession_number": {
            "type": "string",
            "pattern": "^\\d{10}-\\d{2}-\\d{6}$",
            "title": "Accession Number"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/AttachmentInfo"
            },
            "type": "array",
            "title": "Attachments"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "accession_number",
          "total",
          "attachments"
        ],
        "title": "AttachmentsResponse"
      },
...
      "ContentResponse": {
        "properties": {
          "accession_number": {
            "type": "string",
            "pattern": "^\\d{10}-\\d{2}-\\d{6}$",
            "title": "Accession Number"
          },
          "fmt": {
            "type": "string",
            "enum": [
              "markdown",
              "text",
              "html"
            ],
            "title": "Fmt"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "accession_number",
          "fmt",
          "content"
        ],
        "title": "ContentResponse"
      },
...
      "DocumentRef": {
        "properties": {
          "sequence": {
            "type": "string",
            "title": "Sequence"
          },
          "document": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "document_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Type"
          },
          "size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size"
          },
          "ixbrl": {
            "type": "boolean",
            "title": "Ixbrl"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sequence",
          "document",
          "description",
          "document_type",
          "size",
          "ixbrl",
          "url"
        ],
        "title": "DocumentRef"
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "Canonical body for every non-2xx the sidecar emits (HTTPException, mapped\nedgartools/httpx exceptions, and flattened request-validation errors alike)."
      },
...
      "FilingEntity": {
        "properties": {
          "cik": {
            "type": "string",
            "pattern": "^\\d{10}$",
            "title": "Cik"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "form": {
            "type": "string",
            "title": "Form"
          },
          "filing_date": {
            "type": "string",
            "format": "date",
            "title": "Filing Date"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "cik",
          "company",
          "form",
          "filing_date"
        ],
        "title": "FilingEntity",
        "description": "One quarterly-index row for the accession (primary entity first)."
      },
      "FilingEnvelope": {
        "properties": {
          "accession_number": {
            "type": "string",
            "pattern": "^\\d{10}-\\d{2}-\\d{6}$",
            "title": "Accession Number"
          },
          "form": {
            "type": "string",
            "title": "Form"
          },
          "cik": {
            "type": "string",
            "pattern": "^\\d{10}$",
            "title": "Cik"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "filing_date": {
            "type": "string",
            "format": "date",
            "title": "Filing Date"
          },
          "is_multi_entity": {
            "type": "boolean",
            "title": "Is Multi Entity"
          },
          "entities": {
            "items": {
              "$ref": "#/components/schemas/FilingEntity"
            },
            "type": "array",
            "title": "Entities"
          },
          "header": {
            "$ref": "#/components/schemas/FilingHeaderModel"
          },
          "primary_documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentRef"
            },
            "type": "array",
            "title": "Primary Documents"
          },
          "homepage_url": {
            "type": "string",
            "title": "Homepage Url"
          },
          "text_url": {
            "type": "string",
            "title": "Text Url"
          },
          "obj_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Obj Type"
          },
          "data": {
            "type": "null",
            "title": "Data"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "accession_number",
          "form",
          "cik",
          "company",
          "filing_date",
          "is_multi_entity",
          "entities",
          "header",
          "primary_documents",
          "homepage_url",
          "text_url",
          "obj_type",
          "data"
        ],
        "title": "FilingEnvelope"
      },
      "FilingHeaderModel": {
        "properties": {
          "acceptance_datetime": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acceptance Datetime"
          },
          "filing_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filing Date"
          },
          "period_of_report": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Of Report"
          },
          "date_as_of_change": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date As Of Change"
          },
          "document_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Count"
          },
          "items": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Items"
          },
          "filers": {
            "items": {
              "$ref": "#/components/schemas/HeaderFiler"
            },
            "type": "array",
            "title": "Filers"
          },
          "reporting_owners": {
            "items": {
              "$ref": "#/components/schemas/HeaderReportingOwner"
            },
            "type": "array",
            "title": "Reporting Owners"
          },
          "issuer": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HeaderIssuer"
              },
              {
                "type": "null"
              }
            ]
          },
          "subject_companies": {
            "items": {
              "$ref": "#/components/schemas/HeaderSubjectCompany"
            },
            "type": "array",
            "title": "Subject Companies"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "acceptance_datetime",
          "filing_date",
          "period_of_report",
          "date_as_of_change",
          "document_count",
          "items",
          "filers",
          "reporting_owners",
          "issuer",
          "subject_companies"
        ],
        "title": "FilingHeaderModel"
      },
...
      "HeaderFiler": {
        "properties": {
          "company": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CompanyInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "filing_values": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FilingValues"
              },
              {
                "type": "null"
              }
            ]
          },
          "business_address": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Address"
              },
              {
                "type": "null"
              }
            ]
          },
          "mailing_address": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Address"
              },
              {
                "type": "null"
              }
            ]
          },
          "former_names": {
            "items": {
              "$ref": "#/components/schemas/FormerName"
            },
            "type": "array",
            "title": "Former Names"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "company",
          "filing_values",
          "business_address",
          "mailing_address",
          "former_names"
        ],
        "title": "HeaderFiler"
      },
      "HeaderIssuer": {
        "properties": {
          "company": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CompanyInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "business_address": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Address"
              },
              {
                "type": "null"
              }
            ]
          },
          "mailing_address": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Address"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "company",
          "business_address",
          "mailing_address"
        ],
        "title": "HeaderIssuer"
      },
      "HeaderReportingOwner": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "cik": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^\\d{10}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cik"
          },
          "company": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CompanyInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "filing_values": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FilingValues"
              },
              {
                "type": "null"
              }
            ]
          },
          "business_address": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Address"
              },
              {
                "type": "null"
              }
            ]
          },
          "mailing_address": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Address"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "cik",
          "company",
          "filing_values",
          "business_address",
          "mailing_address"
        ],
        "title": "HeaderReportingOwner"
      },
...
      "SectionInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "part": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part"
          },
          "item": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item"
          },
          "detection_method": {
            "type": "string",
            "title": "Detection Method"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "title",
          "part",
          "item",
          "detection_method",
          "confidence",
          "content"
        ],
        "title": "SectionInfo"
      },
      "SectionsResponse": {
        "properties": {
          "accession_number": {
            "type": "string",
            "pattern": "^\\d{10}-\\d{2}-\\d{6}$",
            "title": "Accession Number"
          },
          "fmt": {
            "type": "string",
            "enum": [
              "text",
              "markdown"
            ],
            "title": "Fmt"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/SectionInfo"
            },
            "type": "array",
            "title": "Sections"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "accession_number",
          "fmt",
          "total",
          "sections"
        ],
        "title": "SectionsResponse"
      },
...
    }
  }
}